+++ src/main/java/org/dti/se/miniproject1backend1/inners/models/valueobjects/participant/CreateFeedbackResponse.java
@@ -1,4 +1,4 @@
-package org.dti.se.miniproject1backend1.inners.models.valueobjects.profile;
+package org.dti.se.miniproject1backend1.inners.models.valueobjects.participant;
 
 import lombok.*;
 import lombok.experimental.Accessors;
@@ -12,8 +12,9 @@ import java.util.UUID;
 @AllArgsConstructor
 @NoArgsConstructor
 @Accessors(chain = true)
-public class RetrieveFeedbackResponse extends Model {
+public class CreateFeedbackResponse extends Model {
     private UUID id;
+    private UUID transactionId;
     private Integer rating;
     private String review;
 }
