+++ src/main/java/org/dti/se/miniproject1backend1/inners/models/valueobjects/participant/RetrieveFeedbackResponse.java
@@ -1,4 +1,4 @@
-package org.dti.se.miniproject1backend1.inners.models.valueobjects.profile;
+package org.dti.se.miniproject1backend1.inners.models.valueobjects.participant;
 
 import lombok.*;
 import lombok.experimental.Accessors;
@@ -12,8 +12,8 @@ import java.util.UUID;
 @AllArgsConstructor
 @NoArgsConstructor
 @Accessors(chain = true)
-public class CreateFeedbackRequest extends Model {
-    private UUID transactionId;
+public class RetrieveFeedbackResponse extends Model {
+    private UUID id;
     private Integer rating;
     private String review;
 }
