+++ src/main/java/org/dti/se/miniproject1backend1/inners/usecases/participant/BasicParticipantUseCase.java
@@ -1,13 +1,13 @@
-package org.dti.se.miniproject1backend1.inners.usecases.profile;
+package org.dti.se.miniproject1backend1.inners.usecases.participant;
 
 import org.dti.se.miniproject1backend1.inners.models.entities.Account;
 import org.dti.se.miniproject1backend1.inners.models.entities.Event;
 import org.dti.se.miniproject1backend1.inners.models.entities.Feedback;
 import org.dti.se.miniproject1backend1.inners.models.entities.Transaction;
-import org.dti.se.miniproject1backend1.inners.models.valueobjects.profile.CreateFeedbackRequest;
-import org.dti.se.miniproject1backend1.inners.models.valueobjects.profile.CreateFeedbackResponse;
-import org.dti.se.miniproject1backend1.inners.models.valueobjects.profile.RetrieveAllFeedbackResponse;
-import org.dti.se.miniproject1backend1.inners.models.valueobjects.profile.RetrieveFeedbackResponse;
+import org.dti.se.miniproject1backend1.inners.models.valueobjects.participant.CreateFeedbackRequest;
+import org.dti.se.miniproject1backend1.inners.models.valueobjects.participant.CreateFeedbackResponse;
+import org.dti.se.miniproject1backend1.inners.models.valueobjects.participant.RetrieveAllFeedbackResponse;
+import org.dti.se.miniproject1backend1.inners.models.valueobjects.participant.RetrieveFeedbackResponse;
 import org.dti.se.miniproject1backend1.outers.exceptions.accounts.AccountUnAuthorizedException;
 import org.dti.se.miniproject1backend1.outers.repositories.ones.EventRepository;
 import org.dti.se.miniproject1backend1.outers.repositories.ones.FeedbackRepository;
@@ -21,7 +21,7 @@ import java.util.List;
 import java.util.UUID;
 
 @Service
-public class BasicProfileUseCase {
+public class BasicParticipantUseCase {
     @Autowired
     TransactionRepository transactionRepository;
